refactor : Member도메인 수정, 프로필 이미지 필드를 String이 아닌 Enum으로 수정
