--- aiku/aiku-common/src/main/java/common/domain/MemberProfile.java
@@ -1,19 +1,0 @@
-package common.domain;
-
-import jakarta.persistence.Embeddable;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PACKAGE)
-@Embeddable
-public class MemberProfile {
-
-    private String profileType;
-    private String profileImg;
-    private String profileCharacter;
-    private String profileBackground;
-}
