+++ aiku/aiku-common/src/main/java/common/domain/event/RecommendEvent.java
@@ -1,7 +1,6 @@
 package common.domain.event;
 
-import common.domain.BaseTime;
-import common.domain.Member;
+import common.domain.member.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
