+++ aiku/aiku-common/src/main/java/common/domain/member/MemberProfileBackground.java
@@ -1,0 +1,5 @@
+package common.domain.member;
+
+public enum MemberProfileBackground {
+    RED, GREEN, BLUE
+}
