+++ aiku/aiku-common/src/main/java/common/domain/member/MemberProfileCharacter.java
@@ -1,0 +1,5 @@
+package common.domain.member;
+
+public enum MemberProfileCharacter {
+    C01, C02, C03, C04
+}
