+++ aiku/aiku-common/src/main/java/common/domain/member/MemberProfileType.java
@@ -1,0 +1,5 @@
+package common.domain.member;
+
+public enum MemberProfileType {
+    IMG, CHAR
+}
