+++ aiku/aiku-common/src/test/java/common/domain/TeamTest.java
@@ -1,12 +1,11 @@
 package common.domain;
 
-import org.assertj.core.api.Assertions;
+import common.domain.member.Member;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class TeamTest {
