+++ aiku/aiku-main/src/main/java/aiku_main/dto/TeamMemberResDto.java
@@ -1,6 +1,6 @@
 package aiku_main.dto;
 
-import common.domain.Member;
+import common.domain.member.Member;
 import common.domain.TeamMember;
 import lombok.Getter;
