+++ aiku/aiku-main/src/test/java/aiku_main/integration_test/TeamServiceIntegrationTest.java
@@ -5,12 +5,11 @@ import aiku_main.dto.TeamDetailResDto;
 import aiku_main.dto.TeamMemberResDto;
 import aiku_main.repository.TeamRepository;
 import aiku_main.service.TeamService;
-import common.domain.Member;
+import common.domain.member.Member;
 import common.domain.Team;
 import common.domain.TeamMember;
 import common.exception.NoAuthorityException;
 import jakarta.persistence.EntityManager;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
